requirements.txt
requests
python-dotenv
typed-argument-parser

evaluate
pylcs
accelerate
rouge_score
sacrebleu
cer